refactor(QuestionCard): Move to presenter folder
